% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_nuevo_proyecto.R
\name{nuevo_proyecto}
\alias{nuevo_proyecto}
\title{Crea un nuevo proyecto con la estructura `fisabio`}
\usage{
nuevo_proyecto(
  nombre_proyecto = NULL,
  directorio = NULL,
  git = TRUE,
  renv = TRUE
)
}
\arguments{
\item{nombre_proyecto}{Cadena de caracteres con el nombre a asignar al proyecto. Por
defecto se asigna el nombre "fisabio_se_yyXXXX", donde `yy` es el año y XXXX es el
código correlativo de proyecto para ese año. La función espera que cada proyecto esté
ubicado en un directorio raíz destinado a proyectos vigentes y, en caso de que este
argumento se deje en blanco y no se detecte la estructura que este directorio debería
tener, inicia una nueva cadena de proyecto (p. ej., para el primer proyecto del año
2022: fisabio_se_220001).}

\item{directorio}{Cadena de caracteres indicando la ruta al directorio principal donde
se creará el proyecto. Por defecto se escoge el directorio actual de trabajo,
devuelto por la función [base::getwd()]. Si este argumento no esta vacío y el
directorio indicado no existe, la función tratará de crearlo de forma recursiva.}

\item{git}{Valor lógico (TRUE -opción por defecto- o FALSE), indicando si se desea
generar un repositorio `git` asociado al proyecto.}

\item{renv}{Valor lógico (TRUE -opción por defecto- o FALSE), indicando si se desea
generar un control de versiones de paquetes con [renv::init()].}
}
\value{
La función crea la siguiente estructura de directorios en directorio/nombre_proyecto/:
  * r
  * informes
  * datos
      - brutos
      - procesados
      - documentacion
}
\description{
Crea automáticamente un proyecto siguiendo las directrices comunes de `fisabio`,
empleando una jerarquía de directorios unificada, cargando archivos y funciones comunes
del grupo de colaboradores. Opcionalmente, genera un repositorio `git` y un control de
versiones de paquetes de `R` con `renv`.
}
\details{
La función crea la estructura de directorios a partir del nombre del proyecto
  empleando minúsculas, con independencia de si el usuario ha introducido el argumento
  con mayúsculas.

  Los informes se almacenan, junto con sus figuras y datos de caché (si los hubiera),
  en el directorio informes.
}
\examples{
\dontrun{
library(fisabio)
nuevo_proyecto(nombre_proyecto = "proyecto_europeo_X", directorio = "~/proyectos")
}
}
