% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99_utils.R
\name{printest}
\alias{printest}
\title{Formato de tres números como estimación e IC}
\usage{
printest(
  est = NULL,
  lower = NULL,
  upper = NULL,
  texto = "IC 95 \%:",
  digitos = 2
)
}
\arguments{
\item{est}{Estimación.}

\item{lower}{Nivel inferior del IC.}

\item{upper}{Nivel superior del IC.}

\item{texto}{Texto abreviado para intervalo de confianza. Por defecto \verb{IC 95 \%:}.}

\item{digitos}{Número de dígitos del redondeo. Por defecto 2.}
}
\value{
La función devuelve una cadena de texto con la estructura deseada.
}
\description{
Formatea tres números como \verb{EST (IC XX \%: LOWER; UPPER)} fijando el número de decimales.
}
\examples{
printest(est = 0.95, lower = 0.8, upper = 1.13, texto = "IC 95:", digitos = 2)
}
