% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_fisabio.R
\docType{package}
\name{fisabio}
\alias{fisabio-package}
\alias{fisabio}
\title{fisabio}
\description{
Librería con funciones, datos y utilidades de uso habitual para los miembros
del Servicio de Estudios Estadísticos de FISABIO.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/estadistica-fisabio/fisabio}
  \item Report bugs at \url{https://github.com/estadistica-fisabio/fisabio/issues}
}

}
\author{
\strong{Maintainer}: Carlos Vergara-Hernández \email{carlos.vergara@fisabio.es} (\href{https://orcid.org/0000-0002-7263-856X}{ORCID})

Authors:
\itemize{
  \item Ana Escrig-Pérez \email{ana.escrig@fisabio.es}
}

}
\keyword{internal}
