% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_fisabio.R
\name{fisabio}
\alias{fisabio}
\title{fisabio}
\description{
Librería con funciones, datos y utilidades de uso habitual para los miembros
del Servicio de Estudios Estadísticos de FISABIO.
}
\details{
_PACKAGE
}
