% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_informes.R
\name{informe_pdf}
\alias{informe_pdf}
\title{Funci贸n que devuelve el formato de salida apropiado para generar el informe o
la presentaci贸n en PDF.}
\usage{
informe_pdf(motor = c("xelatex", "pdftex"), ...)
}
\arguments{
\item{motor}{Motor de LaTeX. XeLaTeX por defecto.}

\item{...}{Otros argumentos a pasar.}
}
\value{
Objeto con clase "rmarkdown_output_format".
}
\description{
Esta funci贸n crea una lista con todas las especificaciones necesarias para
crear un PDF siguiendo el formato de la plantilla de `fisabio`. Por defecto se
usa XeLaTeX como motor de LaTeX (mejor trabajo con fuentes).
}
\details{
La funci贸n solo debe utilizarse desde una llamada a
  rmarkdown::render() o dentro de un documento .Rmd
}
